Remove automatic registration delay for rhsmcertd
* Card ID: CCT-67 (main branch)
* Card ID: CCT-69 (backport)

The delay was introduced in early versions of the code when Python was
not reliable for long-running tasks. Since we want to make automatic
registration in both standard and anonymous flow to be as fast as
possible, this delay is removed.

This splay period is now performed by the Python code in case the we
have obtained the anonymous entitlement certificates; before we ask for
the identity certificate.

(Cherry-picked from cb18b7901785eb9413905b0ffd3c9123a3ee3b3d)