CCT-69: Update identity reporting in DNF plugin during autoregistration
* Card ID: CCT-66 (main branch)
* Card ID: CCT-67 (backport)

During the first phase of anonymous cloud registration, the system has
valid entitlement certificates that aren't associated with any consumer.

We shouldn't be reporting missing identity. Previously this was an error
state that was not valid, but with automatic registration it is possible
to have a system consuming content that does not have an identity, for
a transition period before it is part of an anonymous or claimed
organization.

(Cherry-picked from 4820e5f61be6613f3408af3cee849c560ce195c3)